search/requirements.api.txt
# API-specific requirements with chat history support
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0

# Existing search pipeline dependencies
numpy==1.24.3
openai==1.3.0
tiktoken==0.5.1
pymongo==4.6.0

# Additional utilities
python-multipart==0.0.6
aiofiles==23.2.1
